# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI World API",
        "description": "API documentation for AI World application",
        "contact": {
            "email": "support@aiworld.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/roles/create": {
            "post": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Create new Role",
                "operationId": "3b1e91e5f753915dde5425115bf4662b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Role name - Validations: min=3, max=50, unique",
                                        "type": "string"
                                    },
                                    "privileges": {
                                        "description": "Array of privilege IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "is_editable": {
                                        "description": "Whether role is editable (default: true)",
                                        "type": "boolean"
                                    },
                                    "is_active": {
                                        "description": "Whether role is active (default: true)",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role created successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "x_localization": []
                    }
                ]
            }
        },
        "/roles/{id}/details": {
            "get": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Get Role details by ID",
                "operationId": "d94ae89106273f672ad00b3a2aa8fe7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "x_localization": []
                    }
                ]
            }
        },
        "/roles/{id}/update": {
            "post": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Update Role",
                "operationId": "789ddc6e11524107002ec595073d0fd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Role name - Validations: min=3, max=50, unique",
                                        "type": "string"
                                    },
                                    "privileges": {
                                        "description": "Array of privilege IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "is_editable": {
                                        "description": "Whether role is editable",
                                        "type": "boolean"
                                    },
                                    "is_active": {
                                        "description": "Whether role is active",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "x_localization": []
                    }
                ]
            }
        },
        "/roles/{id}/delete": {
            "delete": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Delete Role",
                "operationId": "6888d06dda39ed35d49d4d7a546834b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "x_localization": []
                    }
                ]
            }
        },
        "/roles/{id}/change-status": {
            "post": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Change Role status",
                "operationId": "d54c131672b851e77f9205ee4b4a89dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Validations: 0,1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "x_localization": []
                    }
                ]
            }
        },
        "/roles/active/list": {
            "get": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Get all active roles for dropdown",
                "operationId": "666372461a2eb649db9c7b4d9ff6fc08",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "x_localization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter bearer token",
                "name": "Authorization",
                "scheme": "bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
...
}